The rest of the refactor
