+++ src/prefect/settings/models/server/__init__.py
@@ -1,0 +1,1 @@
+from .root import ServerSettings
